renamed package to extensions
