+++ src/main/java/org/saucedemo/extensions/TestExecutionLifecycle.java
@@ -1,4 +1,4 @@
-package org.saucedemo.testlifecycle;
+package org.saucedemo.extensions;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
